1.  Updated the EPICS, NI GPIB, NI VISA, Struck SIS 1100/3100, libusb, and     Jorway CAMAC drivers for compatibility with the 64-bit version of MX.
2.  Successfully tested EPICS motors and scalers with the 64-bit version of MX.

3.  Added code to 'mx/libMx/mx_motor.c' and 'mx/motor/mdisplay.c' to
    distinguish between hardware and software limits in error reporting.



git-svn-id: http://svn.csrri.iit.edu/mx/trunk@254 42547f41-83fc-0310-9aef-9668e019ab9e

+++ libMx/d_epics_timer.h
@@ -32,7 +32,7 @@ typedef struct {
 	MX_EPICS_PV tp_pv;
 	MX_EPICS_PV vers_pv;
 
-	long num_epics_counters;
+	int num_epics_counters;
 	MX_EPICS_PV *gate_control_pv_array;
 } MX_EPICS_TIMER;
